refactor: rename cofacts-exec to cofactsExec for naming consistency
Match the camelCase pattern used by other lib modules (chatCache,
adkClient) so file names follow a single convention.
