+++ README.md
@@ -119,11 +119,11 @@ The following scripts can also be run using pnpm:
 
 ### GraphQL Codegen
 
-Server-side GraphQL queries are typed via [`@graphql-codegen/client-preset`](https://the-guild.dev/graphql/codegen/docs/guides/react-vue#client-preset). Define queries inline with the `graphql` tag from `@/server/gql` and dispatch them through `cofactsExec` (see `src/lib/cofacts-exec.ts`):
+Server-side GraphQL queries are typed via [`@graphql-codegen/client-preset`](https://the-guild.dev/graphql/codegen/docs/guides/react-vue#client-preset). Define queries inline with the `graphql` tag from `@/server/gql` and dispatch them through `cofactsExec` (see `src/lib/cofactsExec.ts`):
 
 ```ts
 import { graphql } from '@/server/gql'
-import { cofactsExec } from '@/lib/cofacts-exec'
+import { cofactsExec } from '@/lib/cofactsExec'
 
 const GetSomethingDocument = graphql(`
   query GetSomething { ... }
